+++ client/src/Components/home/featuredProducts.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Carousel from 'react-multi-carousel';
 import 'react-multi-carousel/lib/styles.css';
-import data from '../../data';
+import data from '../../../../server/data';
 import CarouselProduct from './carousel';
 
 function FeaturedProducts() {
