+++ client/src/Pages/cart.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import data from '../data';
+import data from '../../../server/data';
 
 function Cart() {
   const product = data.product;
