+++ server/data.js
@@ -1,6 +1,7 @@
 const data = {
   product: [
     {
+      id: 1,
       name: 'Bookinees',
       image: 'Images/women/heels/heels02.png',
       brand: ' Jimmy Choo',
@@ -17,6 +18,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 2,
       name: 'Gucci Marmont',
       image: 'Images/women/heels/heels01.png',
       brand: 'Gucci',
@@ -33,6 +35,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 3,
       name: 'Stuart Weitzman Nudis',
       image: 'Images/women/heels/heels03.png',
       brand: 'Stuart Weitzman',
@@ -49,6 +52,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 4,
       name: 'Gucci Marmont',
       brand: 'Gucci',
       price: '12 000',
@@ -65,6 +69,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 5,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '10000',
@@ -79,6 +84,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 6,
       name: 'Gucci Marmont',
       brand: 'Gucci',
       price: '8000 ',
@@ -93,6 +99,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 7,
       name: 'Gucci Marmont',
       brand: 'Gucci',
       price: '8000 frs',
@@ -107,6 +114,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 8,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -121,6 +129,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 9,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -135,6 +144,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 10,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -149,6 +159,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 11,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -163,6 +174,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 12,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -177,6 +189,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 13,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -191,6 +204,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 14,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -205,6 +219,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 15,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -219,6 +234,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 16,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -233,6 +249,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 17,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -248,6 +265,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 18,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -263,6 +281,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 19,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -277,6 +296,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 20,
       name: 'Gucci Marmont',
       brand: 'Gucci',
       price: '8000 ',
@@ -293,6 +313,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 21,
       name: 'Gucci Marmont',
       brand: 'Gucci',
       price: '8000 frs',
@@ -308,6 +329,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 22,
       name: 'Gucci Marmont',
       brand: 'Gucci',
       price: '8000 frs',
@@ -323,6 +345,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 23,
       name: 'Gucci Marmont',
       brand: 'Gucci',
       price: '8000 frs',
@@ -338,6 +361,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 24,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -353,21 +377,7 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
-      name: 'Jimmy Choo Romy',
-      brand: 'Jimmy Choo',
-      price: '8000 frs',
-      category: 'women',
-      countInStock: 4,
-      rating: 3.5,
-      newArrival: true,
-      featured: false,
-      bestSeller: false,
-      trending: false,
-      numReview: 10,
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
-    },
-    {
+      id: 25,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
@@ -383,6 +393,23 @@ const data = {
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
     },
     {
+      id: 26,
+      name: 'Jimmy Choo Romy',
+      brand: 'Jimmy Choo',
+      price: '8000 frs',
+      category: 'women',
+      countInStock: 4,
+      rating: 3.5,
+      newArrival: true,
+      featured: false,
+      bestSeller: false,
+      trending: false,
+      numReview: 10,
+      description:
+        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor',
+    },
+    {
+      id: 27,
       name: 'Jimmy Choo Romy',
       brand: 'Jimmy Choo',
       price: '8000 frs',
